+++ test/viking/model/reflections/belongs_to_reflection.js
@@ -18,7 +18,8 @@
             plural: 'parents',
             routeKey: 'parents',
             singular: 'parent',
-            collection: 'parents'
+            collection: 'parents',
+            collectionName: 'ParentCollection'
         });
         
         delete Parent;